Scrolling background added
Version 0.12
Fixed an error where menu buttons could be pressed while the game was
running.
Added parallax background(it could be simplified in the future if it
causes lag)
changed framerate clock to something more accurate(the old one was
causing lag every second or so) to one that uses more CPU but causes
less lag
print commands cause noticeable lag but they wil be removed  in the
future.